feat: introduce transaction enums and refactor payment status handling across models and controllers

+++ database/migrations/2026_03_25_110228_create_paystack_transactions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('family_id')->constrained()->cascadeOnDelete();
             $table->string('type');
-            $table->unsignedInteger('amount');
+            $table->unsignedBigInteger('amount');
             $table->string('status')->default('pending');
             $table->json('paystack_response')->nullable();
             $table->json('metadata')->nullable();
